requirement.txt
google-cloud-bigquery
pandas
streamlit
pyyaml
matplotlib
seaborn
plotly


